app/core/gimpchannel.h app/core/gimpitem.c
2003-01-14  Manish Singh  <yosh@gimp.org>

        * app/core/gimpchannel.h
        * app/core/gimpitem.c
        * app/display/gimpnavigationview.c
        * app/gui/paths-dialog.c
        * app/tools/gimphistogramtool.c
        * app/tools/gimpscaletool.c
        * app/widgets/gimplistitem.c
        * libgimp/gimppixelrgn.c
        * libgimpwidgets/gimpunitmenu.c
        * plug-ins/FractalExplorer/Dialogs.c
        * plug-ins/common/aa.c
        * plug-ins/common/despeckle.c
        * plug-ins/common/psd.c
        * plug-ins/common/sharpen.c
        * plug-ins/common/snoise.c
        * plug-ins/common/spread.c
        * plug-ins/ifscompose/ifscompose.c
        * plug-ins/xjt/xjt.c: some minor code cleanup

        * plug-ins/common/csource.c: 64-bit cleanliness

+++ plug-ins/common/csource.c
@@ -505,7 +505,7 @@ save_image (Config *config,
 	       s_uint_8,
 	       config->use_rle ? "rle_" : "");
       if (config->use_rle)
-	fprintf (fp, "%u + 1];\n", img_buffer_end - img_buffer);
+	fprintf (fp, "%u + 1];\n", (guint) (img_buffer_end - img_buffer));
       else
 	fprintf (fp, "%u * %u * %u + 1];\n",
 		 drawable->width,
@@ -584,7 +584,7 @@ save_image (Config *config,
 	       macro_name,
 	       config->use_rle ? "rle_" : "");
       if (config->use_rle)
-	fprintf (fp, "%u] =\n", img_buffer_end - img_buffer);
+	fprintf (fp, "%u] =\n", (guint) (img_buffer_end - img_buffer));
       else
 	fprintf (fp, "%u * %u * %u + 1] =\n",
 		 drawable->width,
